requirements3.txt
absl-py==2.3.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
attrs==25.3.0
backoff==2.2.1
blinker==1.8.2
cachetools==5.5.2
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
contourpy==1.1.1
cycler==0.10.0
dataclasses-json==0.6.7
defusedxml==0.7.1
exceptiongroup==1.3.0
fastapi==0.121.0
filelock==3.16.1
filetype==1.2.0
Flask==2.3.3
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-auth-oauthlib==1.0.0
grpcio==1.70.0
h11==0.16.0
httptools==0.6.4
idna==2.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inference-sdk==0.37.1
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
llvmlite==0.41.1
Markdown==3.7
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.7.5
mpmath==1.3.0
multidict==6.1.0
mypy_extensions==1.1.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
oauthlib==3.3.1
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opencv-python-headless==4.8.0.74
packaging==25.0
pandas==2.0.3
Pillow==9.5.0
polars==1.8.2
propcache==0.2.0
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
pymongo>=4.6.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
roboflow==1.1.17
rsa==4.9.1
scipy==1.10.0
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.44.0
supervision==0.25.1
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.189
ultralytics-thop==2.0.16
urllib3==2.2.3
uvicorn==0.33.0
watchfiles==0.24.0
websockets==13.1
Werkzeug==2.3.7
yarl==1.15.2
zipp==3.20.2
